:recycle: Rename .ts data model folder for ExternalTaskApi requests
