+++ typescript/src/data_models/external_task/index.ts
@@ -1,4 +1,4 @@
-export * from './request_post_payloads/index';
+export * from './requests/index';
 export * from './results/index';
 
 export * from './external_task';
