+++ typescript/src/data_models/external_task/requests/fetch_and_lock_request_payload.ts
