+++ typescript/src/data_models/external_task/requests/finish_external_task_request_payload.ts
