+++ typescript/src/data_models/external_task/requests/handle_service_error_request_payload.ts
